feat: add EventBody component
